+++ backend/src/test/java/reviewme/template/service/mapper/TemplateMapperTest.java
@@ -1,4 +1,4 @@
-package reviewme.template.service;
+package reviewme.template.service.mapper;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -30,7 +30,6 @@ import reviewme.template.repository.TemplateRepository;
 import reviewme.template.service.dto.response.QuestionResponse;
 import reviewme.template.service.dto.response.SectionResponse;
 import reviewme.template.service.dto.response.TemplateResponse;
-import reviewme.template.service.module.TemplateMapper;
 
 @ServiceTest
 class TemplateMapperTest {
